==> 01_Extraccion_Datos.ipynb <==
# 1. Importar librerías
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
import pandas as pd
import numpy as np
from selenium.webdriver.common.by import By
import os
import time
import requests

# 2. Utilizar Sellenium para descargar el archivo
url = "https://www.datosabiertos.gob.pe/dataset/postulantes-al-concurso-de-admisi%C3%B3n-de-la-universidad-nacional-de-ingenier%C3%ADa-uni"
driver  = webdriver.Chrome( )
driver.get(url)
driver.maximize_window()
wait = WebDriverWait(driver, 15)
download_button = wait.until(EC.element_to_be_clickable((By.XPATH, '/html/body/div[4]/div/div/section/div/div/div/div/div[2]/div/div/div/article/div/div[3]/div/div/ul/li[1]/div/span/a[2]')))
download_button.click()

# 3. Leer el archivo descargado
download_dir = os.path.expanduser('~/Downloads')
files = os.listdir(download_dir)
paths = [os.path.join(download_dir, basename) for basename in files if basename.lower().endswith(('.csv'))]
latest_file = max(paths, key=os.path.getctime)
df = pd.read_csv(latest_file)
df.head()

# 4. Exportamos los datos sin procesar
data_folder = 'data'
if not os.path.exists(data_folder):
    os.makedirs('data')

output_file = os.path.join(data_folder, 'data_raw.csv')
df.to_csv(output_file, index=False)
