==> 02_Preprocesamiento.ipynb <==
# Importamos librerías
import pandas as pd
import numpy as  np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
import unicodedata
warnings.filterwarnings('ignore')

df = pd.read_csv('data/data_raw.csv')
df.head()

# 1. Creamos una copia para editar el dataframe
df_1 = df.copy()

df_1.columns

# 2.1 Cambiamos el nombre de las columnas para evitar que haya columnas repetidas
df_1['COLEGIO_DEPA'] = df_1['COLEGIO_DEPA'].replace({
    'ÁNCASH': 'ANCASH',
    'JUNÍN': 'JUNIN',
    'HUÁNUCO': 'HUANUCO',
    'SAN MARTÍN': 'SAN MARTIN',
    'APURÍMAC': 'APURIMAC'
})

## Verificamos los valores únicos de la columna de Departamentos
df_1['COLEGIO_DEPA'].unique()

# 2.2 Cambiamos el nombre de las columnas para evitar que haya columnas repetidas
df_1['COLEGIO_PROV'] = df_1['COLEGIO_PROV'].replace({
    'PROV. CONST. DEL CALLAO': 'PROV CONST DEL CALLAO',
    'HUÁNUCO': 'HUANUCO',
    'SÁNCHEZ CARRIÓN': 'SANCHEZ CARRION',
    'CAÑETE': 'CANETE',
    'SAN ROMÁN': 'SAN ROMAN',
    'LA CONVENCIÓN': 'LA CONVENCION',
    'ASUNCIÓN': 'ASUNCION',
    'OYÓN': 'OYON',
    'HUAMALÍES': 'HUAMALIES',
    'AZÁNGARO': 'AZANGARO',
    'JAÉN': 'JAEN',
    'BONGARÁ': 'BONGARA',
    'MORROPÓN': 'MORROPON',
    'CARLOS FERMÍN FITZCARRALD': 'CARLOS FERMIN FITZCARRALD',
    'DANIEL ALCIDES CARRIÓN': 'DANIEL ALCIDES CARRION',
    'VILCAS HUAMÁN': 'VILCAS HUAMAN',
    'MARISCAL CÁCERES': 'MARISCAL CACERES',
    'VIRÚ': 'VIRU',
    'HUANCANÉ': 'HUANCANE',
    'PÁUCAR DEL SARA SARA': 'PAUCAR DEL SARA SARA',
    'LA UNIÓN': 'LA UNION',
    'JULCÁN': 'JULCAN',
    'ANTONIO RAIMONDI': 'ANTONIO RAYMONDI',
    'RODRÍGUEZ DE MENDOZA': 'RODRIGUEZ DE MENDOZA',
    'MARAÑÓN': 'MARANON',
    'CONTUMAZÁ': 'CONTUMAZA',
    'ITALIA.': 'ITALIA',
    'CARABOBO.': 'CARABOBO',
    'MARISCAL RAMÓN CASTILLA': 'MARISCAL RAMON CASTILLA',
    'VÍCTOR FAJARDO': 'VICTOR FAJARDO',
    'SÁNCHEZ CARRIÓN': 'SANCHEZ CARRION',
    'LAURICOCHA.': 'LAURICOCHA',
    'CALGARY.': 'CALGARY',
    'CONDORCANQUI.': 'CONDORCANQUI',
    'ECUADOR.': 'ECUADOR',
    'SUCRE.': 'SUCRE',
    'ATALAYA.': 'ATALAYA',
    'BOLÍVAR': 'BOLIVAR',
    'CASTROVIRREYNA.': 'CASTROVIRREYNA',
    'URUBAMBA.': 'URUBAMBA',
    'HUAROCHIRI.': 'HUAROCHIRI',
    'CAMANA.': 'CAMANA',
    'LAMBAYEQUE.': 'LAMBAYEQUE',
    'FERREÑAFE.': 'FERREÑAFE',
    'CARAVELÍ.': 'CARAVELI',
    'CHEPÉN': 'CHEPEN',
    'LA CONVENCION.': 'LA CONVENCION',
    'CARAVELI.': 'CARAVELI',
    'MAYNAS.': 'MAYNAS',
    'HUALGAYOC.': 'HUALGAYOC',
    'LORETO.': 'LORETO',
    'CUTERVO.': 'CUTERVO',
    'CAMANÁ.': 'CAMANA',
    'QUISPICANCHI.': 'QUISPICANCHI',
    'YUNGAY.': 'YUNGAY',
    'TRUJILLO.': 'TRUJILLO',
    'LAMPA.': 'LAMPA',
    'ESPINAR.': 'ESPINAR',
    'TAYACAJA.': 'TAYACAJA',
    'TARMA.': 'TARMA',
    'BOLOGNESI.': 'BOLOGNESI',
    'CHANCHAMAYO.': 'CHANCHAMAYO',
    'SULLANA.': 'SULLANA',
    'HUANCAYO.': 'HUANCAYO',
    'HUARI.': 'HUARI',
    'CHICLAYO.': 'CHICLAYO',
    'TOCACHE.': 'TOCACHE',
    'PISCO.': 'PISCO',
    'CHANCHAMAYO.': 'CHANCHAMAYO',
    'PROV CONST DEL CALLAO.': 'PROV CONST DEL CALLAO',
    'HUAROCHIRÍ': 'HUAROCHIRI',
    'VILCAS HUAMAN': 'VILCAS HUAMAN',
    'CAMANÁ': 'CAMANA',
    'LA UNIÒN': 'LA UNION',
    'PÀUCAR DEL SARA SARA': 'PAUCAR DEL SARA SARA',
    'SAN MARTÍN': 'SAN MARTIN',
    'CARAVELÍ': 'CARAVELI',
    'CELENDÍN': 'CELENDIN',
    'HUAYTARÁ': 'HUAYTARA',
    'CONCEPCIÓN': 'CONCEPCION',
    'VILCAS HUAMAN': 'VILCAS HUAMÁN',
    'LA UNION': 'LA UNIÒN',
    'JUNIN': 'JUNÍN',
    'PAUCAR DEL SARA SARA' : 'PÀUCAR DEL SARA SARA'

})

## Verificamos los valores únicos de la columna de Provincias
df_1['COLEGIO_PROV'].unique()

# 2.3 Cambiamos el nombre de las columnas para evitar que haya columnas repetidas
## Debido a la gran cantidad de distritos, utilizamos una función para reemplazar tildes y caracteres especiales

def eliminar_tildes(texto):
    if isinstance(texto, str):
        return ''.join(c for c in unicodedata.normalize('NFD', texto) if unicodedata.category(c) != 'Mn')
    else:
        return texto 

## Reemplazar los nombres de distritos utilizando la función
df_1['COLEGIO_DIST'] = df_1['COLEGIO_DIST'].apply(eliminar_tildes)

## Verificamos los valores únicos de la columna de Distritos
df_1['COLEGIO_DIST'].unique()

# 4. Crear una nueva columna para las modalidades resumidas en solo 13 categorías, para una mejor visualización en el Dashboard
## Pasamos de 45 modalidades a 13, para poder graficarlo mejor en el Dashboard de Power BI

print(f'Tenemos {len(df_1.MODALIDAD.unique())} modalidades en la columna original')

modalidades = {
    'ORDINARIO': 'Ingreso Ordinario',

    'EXTRAORDINARIO1 - DEPORTISTAS CALIFICADOS DE ALTO NIVEL( Iniciar estudios)': 'Ingreso por Deportistas Calificados',
    'EXTRAORDINARIO 1 - DEPORTISTAS CALIFICADOS DE ALTO NIVEL( Iniciar estudios)': 'Ingreso por Deportistas Calificados',
    'EXTRAORDINARIO - DEPORTISTA CALIFICADO DE ALTO NIVEL( Iniciar estudios)': 'Ingreso por Deportistas Calificados',

    'EXTRAORDINARIO1 - CONVENIO ANDRES BELLO (iniciar estudios)': 'Ingreso por Convenios',
    'EXTRAORDINARIO 1 - CONVENIO ANDRÉS BELLO (iniciar estudios)': 'Ingreso por Convenios',
    'EXTRAORDINARIO1 - CONVENIO ANDRES BELLO (continuar estudios)': 'Ingreso por Convenios',
    'EXTRAORDINARIO 1 - CONVENIO ANDRÉS BELLO (continuar estudios)': 'Ingreso por Convenios',
    'EXTRAORDINARIO 1 - CONVENIO DIPLOMATICO': 'Ingreso por Convenios',

    'EXTRAORDINARIO - DOS PRIMEROS ALUMNOS': 'Ingreso por Excelencia Académica',
    'EXTRAORDINARIO 1 - DOS PRIMEROS ALUMNOS': 'Ingreso por Excelencia Académica',
    'EXTRAORDINARIO1 - DOS PRIMEROS ALUMNOS': 'Ingreso por Excelencia Académica',

    'EXTRAORDINARIO - VÍCTIMA DEL TERRORISMO (iniciar estudios)': 'Ingreso por Víctimas de Terrorismo',
    'EXTRAORDINARIO 1 - VÍCTIMA DEL TERRORISMO (iniciar estudios)': 'Ingreso por Víctimas de Terrorismo',
    'EXTRAORDINARIO1 - VICTIMAS DEL TERRORISMO (iniciar estudios)': 'Ingreso por Víctimas de Terrorismo',
    'EXTRAORDINARIO 1 - VICTIMAS DEL TERRORISMO (continuar estudios)': 'Ingreso por Víctimas de Terrorismo',
    'EXTRAORDINARIO - VICTIMA DEL TERRORISMO (continuar estudios)': 'Ingreso por Víctimas de Terrorismo',
    'EXTRAORDINARIO1 - VICTIMAS DEL TERRORISMO (continuar estudios)': 'Ingreso por Víctimas de Terrorismo',

    'EXTRAORDINARIO - TRASLADO EXTERNO': 'Ingreso por Traslado Externo',
    'EXTRAORDINARIO1 - TRASLADO EXTERNO': 'Ingreso por Traslado Externo',
    'EXTRAORDINARIO 1 - TRASLADO EXTERNO': 'Ingreso por Traslado Externo',
    'EXTRAORDINARIO 1 - TRASLADO EXTERNO PARA ESTUDIANTES PROVENIENTES DE UNIVERSIDADES NO LICENCIADAS': 'Ingreso por Traslado Externo',
    'EXTRAORDINARIO1 - TRASLADO EXTERNO PARA ESTUDIANTES PROVENIENTES DE UNIVERSIDADES NO LICENCIADAS': 'Ingreso por Traslado Externo',
    'EXTRAORDINARIO - TRASLADO EXTERNO PARA ESTUDIANTES PROVENIENTES DE UNIVERSIDADES NO LICENCIADAS': 'Ingreso por Traslado Externo',

    'EXTRAORDINARIO - TITULADOS O GRADUADOS UNI': 'Ingreso por Titulados o Graduados',
    'EXTRAORDINARIO1 - TITULADOS O GRADUADOS': 'Ingreso por Titulados o Graduados',
    'EXTRAORDINARIO1 - TITULADOS O GRADUADOS UNI': 'Ingreso por Titulados o Graduados',
    'EXTRAORDINARIO 1 - TITULADOS O GRADUADOS EN OTRA UNIVERSIDAD': 'Ingreso por Titulados o Graduados',
    'EXTRAORDINARIO - TITULADOS O GRADUADOS EN OTRA UNIVERSIDAD': 'Ingreso por Titulados o Graduados',
    'EXTRAORDINARIO 1 - TITULADO O GRADUADO UNI': 'Ingreso por Titulados o Graduados',

    'EXTRAORDINARIO - DIPLOMADO CON BACHILLERATO INTERNACIONAL': 'Ingreso por Bachillerato Internacional',
    'EXTRAORDINARIO 1 - DIPLOMADOS CON BACHILLERATO': 'Ingreso por Bachillerato Internacional',
    'EXTRAORDINARIO1 - DIPLOMADOS CON BACHILLERATO INTERNACIONAL': 'Ingreso por Bachillerato Internacional',
    'EXTRAORDINARIO 1 - DIPLOMADOS CON BACHILLERATO': 'Ingreso por Bachillerato Internacional',

    'EXTRAORDINARIO - PERSONA CON DISCAPACIDAD (iniciar estudios)': 'Ingreso por Discapacidad',
    'EXTRAORDINARIO 1 - PERSONAS CON DISCAPACIDAD (iniciar estudios)': 'Ingreso por Discapacidad',
    'EXTRAORDINARIO1 - PERSONAS CON DISCAPACIDAD (iniciar estudios)': 'Ingreso por Discapacidad',

    'EXTRAORDINARIO INGRESO DIRECTO CEPRE-UNI': 'Ingreso Directo CEPRE',
    'EXTRAORDINARIO1 - INGRESO DIRECTO CEPRE': 'Ingreso Directo CEPRE',
    'EXTRAORDINARIO INGRESO DIRECTO CEPRE-UNI O CEPRE-UNI INTENSIVO': 'Ingreso Directo CEPRE',
    'EXTRAORDINARIO 2 – INGRESO DIRECTO CEPRE': 'Ingreso Directo CEPRE',
    'EXTRAORDINARIO2 – INGRESO DIRECTO CEPRE': 'Ingreso Directo CEPRE',

    'TALENTO BECA 18': 'Ingreso por Beca',
    'EXTRAORDINARIO2 - TALENTO BECA 18': 'Ingreso por Beca',

    'INGRESO ESCOLAR NACIONAL': 'Ingreso Escolar Nacional',

    'INTERESADO': 'Otros'
}

df_1['MODALIDAD_RESUM'] = df_1['MODALIDAD'].map(modalidades)

df_1['MODALIDAD_RESUM'] = df_1['MODALIDAD_RESUM'].fillna('Otros')

## Creamos MODALIDADES_RESUM donde se resumen las modalidades a solo 13
print(f'Tenemos {len(df_1.MODALIDAD_RESUM.unique())} modalidades en la columna resumida. Estas son: {df_1.MODALIDAD_RESUM.unique()}')


# 5. Creamos una nueva variable categórica según la califación final de los estudiantes
## Menos de 11: DESAPROBADO
## Entre 11 y 15: APROBADO
## Más de 15: SOBRESALIENTE
labels = ["DESAPROBADO", "APROBADO", "SOBRESALIENTE"]
bins = [0, 11, 15, df_1["CALIF_FINAL"].max()]

## Creamos CALIF_CATEGORICO
df_1["CALIF_CATEGORICO"] = pd.cut(df_1["CALIF_FINAL"],
                                  bins=bins,
                                  labels=labels)


# 6. Analizamos el nuevo dataframe y lo exportamos a Excel
df_1.head()

## Lo exportamos a Excel
df_1.to_excel('data/data_preprocessed.xlsx', index=False)
