==> 04_Tasa_Ingresantes.ipynb <==
# Import libraries
import pandas as pd
import numpy as  np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')

# 1. Cargamos e inspeccionamos los datos

df = pd.read_excel('data/data_preprocessed.xlsx')
df.head()

df.columns

# 2. Creamos un nuevo dataframe que indique la tasa de ingreso por departamento

## Filtramos las observaciones donde INGRESO es 'SI'
df_si = df[df['INGRESO'] == 'SI']

## Contamos las observaciones que tienen 'SI' en INGRESO por departamento
ingresos_por_departamento = df_si.groupby('COLEGIO_DEPA').size()

## Contamos el total de observaciones por departamento
total_por_departamento = df.groupby('COLEGIO_DEPA').size()

## Calculamos la tasa de ingreso (ingresos / total por departamento)
tasa_ingreso = ingresos_por_departamento / total_por_departamento

## Convertimos el resultado a un DataFrame y filtramos los países (no son departamentos)
tasa_ingreso_df = tasa_ingreso.reset_index(name='Tasa_Ingreso')

tasa_ingreso_df = tasa_ingreso_df.dropna()

tasa_ingreso_df = tasa_ingreso_df[~tasa_ingreso_df['COLEGIO_DEPA'].isin(['ARGENTINA', 'ESPAÑA'])]

tasa_ingreso_df

# 3. Ordenamos el DataFrame por Tasa_Ingreso de forma descendente
tasa_ingreso_df = tasa_ingreso_df.sort_values(by='Tasa_Ingreso', ascending=False)
tasa_ingreso_df

# 4. Guardamos el DataFrame en un archivo Excel
tasa_ingreso_df.to_excel('data/tasa_ingreso.xlsx', index=False)
